Transformed all caches to XWikiCache Move virtual wiki cache to XWikiCache, limited by default to 1000 items Allow caches to be completely deleted on flushAll Allow clustered cache to only use one jgroups notification bus Updated oscache to a lightly modified version of oscache
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@1098 f329d543-caf0-0310-9063-dda96c69346f

+++ src/main/java/com/xpn/xwiki/XWiki.java
@@ -1411,8 +1411,10 @@ public class XWiki implements XWikiDocChangeNotificationInterface, XWikiInterfac
         // We need to flush the virtual wiki list
         virtualWikiList = new ArrayList();
         // We need to flush the server Cache
-        virtualWikiMap.flushAll();
-        virtualWikiMap = null;
+        if (virtualWikiMap!=null) {
+            virtualWikiMap.flushAll();
+            virtualWikiMap = null;
+        }
 
         // We need to flush the group service cache
         if (groupService != null)
